Add new internal API for periodic system status report
The system status is reported by default on each ntp time synchronisation,
which by default is on every hour. However, for debug purposes, it is
useful to report the status more often. Currently, the is such logic in
the code, which can be enabled at build time. It is useful to change
that behaviour runtime, not at build time. Added new API and logic for
that:
  system_set_periodic_log_ms()

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>