Split status logs
There is an internal logic to dump the device status on given event.
Each module has a status callback, which is called in that case.
Currently all those callbacks are called at once, in a single loop. As
the logic gets more complex and count of the modules increases, that
status loop can take a lot time to execute. This could cause a WD hit,
or overflow of internal log or network buffer. To solve that problem, the
status log logic is rewritten, so it is executed asynchronous instead in
a single blocking loop.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>