Enhance the params_crypt logic to add missing parameters.
Usually, not all of the configurable user parameters are used in an
application. In order to compile the code, unused parameters must be
defined as empty. The existing approach is to define them in the parameter
file of each application, which complicates the logic when new parameter
is introduced.
This change extends params_crypt.sh to automatically add empty defines
of parameters, which are missing from the application parameter file.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>