Refactor commands handlers
The existing logic for registering and executing custom commands is
designed to be used with the web server. The server runs registered
command when an http request is received. In order to reuse that logic
with MQTT, the code is refactored. A new command execution context is
introduced, which allows to add support for more command triggers.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>