Introduce MQTT custom commands
Added logic to subscribe for specific MQTT topic and execute custom
commands on receiving data on that topic. MQTT broker uses the already
established TCP connection for communication with the client. Using this
channel for sending commands to device is very useful when there is no
direct link to the device.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>