Infrastructure for system modules.
Current design of the project consist common library and application
logic. In order to add support for new IoT device, the logic for it
should be implemented in an application. If more than one application
needs to use the same IoT device, the code has to be copy/pasted.
To simplify support and reusability of new IoT devices, new
infrastructure for modules is introduced. A module implements support
for an IoT device - initialisation, interaction, web and MQTT support.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>