Lock the IP stack when WiFi APIs are called.
The raspberry IP stack is not thread safe, that's why it has a lock.
Every call to network related API must be guarded by that lock. Even
thought the code does not run in parallel threads, part of it is
executed in interrupt context.
Add logic to get / release that lock when a WiFi API is called.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>