Check for parameter length before decoding
If a configuration parameter is not defined in params.txt file, its size
should be 0. Check for that size in the context of the decoding
function, to ensure that these missing parameters are handled correctly.
The current approach is that check to be performed before calling the
decoding function. Moving the check in the decoding logic makes the code
more robust and consistent.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>