Introduce common engine for managing the commands.
Currently the logic for registering, parsing and executing commands is
in MQTT and WEB modules, as both of them are responsible for receiving
user commands. There is a lot of duplicated code. This approach is not
flexible and robust. When another command module is added, the same
code should be copied and modified.
Introduce a new "command engine" module, responsible for user commands.
Removed all duplicated code and simplified MQTT and WEB implementation.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>