BMS: Do not send notification on first run
When the device boots, the first calculation of the battery state
triggers new web hook notification. This can be misleading, as the
battery state is not actually changed, the change have been happened
before the device boot. Added logic to not send web hook notification in
that case, when the battery state is calculated for the first time.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>