MQTT: log status on chunks.
The MQTT status log can be huge, as it dumps all registered MQTT
sensors. Printing huge data at once can overload the TCP stack of the
pico, if the log is printed on a HTTP session. Added new logic to print
the MQTT status on small chunks instead of once.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>