SHT20 sensor: add support for power pin
SHT20 sensor is not stable in long run, needs hardware reset to recover.
It has 3.3V power supply with very low power consumption, which makes it
possible to use regular GPIO pin for power supply the sensor. Added
logic for that - optional usage of GPIO pin for powering the sensor.
Added logic for hard reset, if software reset cannot recover the sensor.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>