Flow sensor: accumulate total flow for given interval
Currently the flow logic measures the flow and accumulates the total
measurements until a reset command is executed. All measurements are
sent on MQTT on every change. That logic depends on raspberry uptime -
if the device reboots, the accumulated total flow is lost.
Added logic to overcome that limitation: optional FLOW_ACC_SEC
parameter. If that is defined, the flow is accumulated in FLOW_ACC_SEC
seconds window. At the end of that interval, the total is sent on MQTT
and the value is cleared. In case of device reboot, only the
measurements of the last window are lost.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>