Switch from RTC to AON time APIs
The rtc hardware and APIs are available only on Pico and Pico W boards.
The next generation Pico 2 and Pico 2 W does not have rtc. The SDK
provides abstraction layer to handle that - AON timer. Refactor the code
to use AON APIs instead of RTC, to simplify support for next generation
Pico 2.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>