MQTT client clean-up
The lock logic of MQTT client is not used, as the application runs in
single thread. Removed that dead code.
Deleting of a running MQTT client causes random crashed and reboots, the
MQTT library does not handle correctly that logic. Instead of deleting,
just disconnect the client on an  error or timeout.

Signed-off-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>